// Frontend/PurchasingManagement/src/app/models/workflow-list-model/workflow-list-model.ts
export class WorkflowListModel {
  id: any | undefined;
  definitionId: any | undefined;
  workflowName: any | undefined;
  displayName: any | undefined;
  description: any | undefined;
  version: any | undefined;
  persistenceBehavior: any | undefined;
  isPublished: any | undefined;
  isLatest: any | undefined;
}